+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- don't write debug “allitems.json” file
+
 ## [0.8.0] - 2024-07-07
 
 ### Added
@@ -34,7 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - (No changelog yet)
 
-[unreleased]: https://github.com/graphinx/graphinx/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/graphinx/graphinx/compare/v0.8.0...HEAD
 [0.8.0]: https://github.com/graphinx/graphinx/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/graphinx/graphinx/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/graphinx/graphinx/releases/tag/v0.6.0